[MIG] sale_order_general_discount: Migration to 13.0
[MIG] Migrated sale_order_general_discount to v13.

[IMP] Improved code as per request.

[IMP][13.0] sale_order_general_discount: rebase with OCA/13.0 + apply pre-commit

[UPD] sale_order_general_discount: update view to display percentage correctly
