+++ sale_order_general_discount/i18n/it.po
@@ -4,45 +4,44 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 11.0\n"
+"Project-Id-Version: Odoo Server 12.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2018-07-31 09:08+0200\n"
-"PO-Revision-Date: 2020-10-26 12:08+0000\n"
-"Last-Translator: Daniel Martinez Vila <daniel.martinez@qubiq.es>\n"
-"Language-Team: \n"
-"Language: es\n"
+"PO-Revision-Date: 2020-03-18 17:13+0000\n"
+"Last-Translator: Francesco Foresti <francesco.foresti@ooops404.com>\n"
+"Language-Team: none\n"
+"Language: it\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
+"Content-Transfer-Encoding: \n"
 "Plural-Forms: nplurals=2; plural=n != 1;\n"
 "X-Generator: Weblate 3.10\n"
 
 #. module: sale_order_general_discount
 #: model:ir.model,name:sale_order_general_discount.model_res_partner
 msgid "Contact"
-msgstr "Contacto"
+msgstr "Contatto"
 
 #. module: sale_order_general_discount
 #: model:ir.model.fields,field_description:sale_order_general_discount.field_res_partner__sale_discount
 #: model:ir.model.fields,field_description:sale_order_general_discount.field_res_users__sale_discount
+#, fuzzy
 msgid "Discount"
-msgstr "Descuento"
+msgstr "Sconto (%)"
 
 #. module: sale_order_general_discount
 #: model:ir.model.fields,field_description:sale_order_general_discount.field_sale_order__general_discount
 #: model:ir.model.fields,field_description:sale_order_general_discount.field_sale_order_line__discount
 msgid "Discount (%)"
-msgstr "Descuento (%)"
+msgstr "Sconto (%)"
 
 #. module: sale_order_general_discount
 #: model:ir.model,name:sale_order_general_discount.model_sale_order
+#, fuzzy
 msgid "Sales Order"
-msgstr "Pedido de venta"
+msgstr "Ordine di Vendita"
 
 #. module: sale_order_general_discount
 #: model:ir.model,name:sale_order_general_discount.model_sale_order_line
+#, fuzzy
 msgid "Sales Order Line"
-msgstr "Línea Pedido de Venta"
-
-#~ msgid "Quotation"
-#~ msgstr "Presupuesto"
+msgstr "Ordine di Vendita"
