+++ sale_order_general_discount/models/__init__.py
@@ -1,3 +1,4 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 from . import res_partner
-from . import sale
+from . import sale_order
+from . import sale_order_line
